En teoría está todo pero el split esta tope de mal

+++ ft_strnstr.c
@@ -6,16 +6,16 @@
 /*   By: aelias-s <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/13 12:15:05 by aelias-s          #+#    #+#             */
-/*   Updated: 2023/03/09 16:29:01 by aelias-s         ###   ########.fr       */
+/*   Updated: 2023/03/18 13:15:57 by aelias-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *dest, const char *src, unsigned int len)
+char	*ft_strnstr(const char *dest, const char *src, size_t len)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	i;
+	size_t	j;
 
 	if (src[0] == '\0')
 		return ((char *)dest);
